Use number of volunteers in chat in web socket messages.
- Update getActiveSubjects in EndpointUtils to use activeVolunteerIDs directly in query.
- Use Set instead of List in ActiveSubjectsMessage to avoid duplicates.
- Create method to sum total number of volunteers in a chat room, and add this number to web sockets messages sent went volunteers enter and leave the chat.

+++ src/main/java/no/capraconsulting/chatmessages/RoomMessage.java
@@ -12,7 +12,7 @@ public class RoomMessage extends Message {
     private List<TextMessage> chatHistory;
     private String talkyID;
     protected String author;
-    private Integer volunteerCount;
+    private long volunteerCount;
 
     public String getVolName() {
         return volName;
@@ -41,7 +41,7 @@ public class RoomMessage extends Message {
     public void setAuthor(String author) { this.author = author;}
     public String getAuthor(){ return author; }
 
-    public Integer getVolunteerCount() {
+    public long getVolunteerCount() {
         return volunteerCount;
     }
 
@@ -53,7 +53,7 @@ public class RoomMessage extends Message {
         private String talkyID;
         private StudentInfo studentInfo;
         private List<TextMessage> chatHistory;
-        private Integer volunteerCount;
+        private long volunteerCount;
 
         public T withStudentID(String studentID) {
             this.studentID = studentID;
@@ -85,7 +85,7 @@ public class RoomMessage extends Message {
             return self();
         }
 
-        public T withVolunteerCount(Integer volunteerCount) {
+        public T withVolunteerCount(long volunteerCount) {
             this.volunteerCount = volunteerCount;
             return self();
         }
